Break up schema
